+++ app/models/product.rb
@@ -64,6 +64,8 @@ class Product < ApplicationRecord
       created_at
       description
       discontinued
+      discontinued_day
+      discontinued_month
       discontinued_year
       id
       name
@@ -96,16 +98,13 @@ class Product < ApplicationRecord
   end
 
   def release_date
-    return nil if release_year.nil?
-    return release_year.to_s if release_month.nil?
+    formatted_date(release_day, release_month, release_year)
+  end
 
-    formatted_month = release_month.to_s.rjust(2, '0')
+  def discontinued_date
+    return unless discontinued?
 
-    return "#{formatted_month}/#{release_year}" if release_day.nil?
-
-    formatted_day = release_day.to_s.rjust(2, '0')
-
-    "#{formatted_day}/#{formatted_month}/#{release_year}"
+    formatted_date(discontinued_day, discontinued_month, discontinued_year)
   end
 
   def price_set
@@ -120,5 +119,20 @@ class Product < ApplicationRecord
     return if description.blank?
 
     Redcarpet::Markdown.new(Redcarpet::Render::HTML.new).render(description)
+  end
+
+  private
+
+  def formatted_date(day, month, year)
+    return nil if year.nil?
+    return year.to_s if month.nil?
+
+    formatted_month = month.to_s.rjust(2, '0')
+
+    return "#{formatted_month}/#{year}" if day.nil?
+
+    formatted_day = day.to_s.rjust(2, '0')
+
+    "#{formatted_day}/#{formatted_month}/#{year}"
   end
 end