+++ db/migrate/20240229062256_add_discontinued_day_and_month_to_product.rb
@@ -1,0 +1,6 @@
+class AddDiscontinuedDayAndMonthToProduct < ActiveRecord::Migration[7.1]
+  def change
+    add_column :products, :discontinued_month, :integer
+    add_column :products, :discontinued_day, :integer
+  end
+end