+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.1].define(version: 2024_02_29_051242) do
+ActiveRecord::Schema[7.1].define(version: 2024_02_29_062256) do
   create_schema "heroku_ext"
 
   # These are extensions that must be enabled in order to support this database
@@ -160,6 +160,8 @@ ActiveRecord::Schema[7.1].define(version: 2024_02_29_051242) do
     t.decimal "price", precision: 12, scale: 4
     t.string "price_currency"
     t.integer "discontinued_year"
+    t.integer "discontinued_month"
+    t.integer "discontinued_day"
     t.index "\"left\"(lower((name)::text), 1)", name: "index_products_name_prefix"
     t.index ["custom_attributes"], name: "index_products_on_custom_attributes", using: :gin
     t.index ["name", "brand_id"], name: "index_products_on_name_and_brand_id", unique: true